@@ -12,27 +12,27 @@
         "url": "https://github.com/krausest/js-framework-benchmark.git"
     },
     "scripts": {
-        "build-dev": "webpack -w -d",
-        "build-prod": "webpack -p"
+        "build-dev": "webpack -w --mode development",
+        "build-prod": "webpack --mode production"
     },
     "dependencies": {
-        "whatsup": "1.1.1",
-        "@whatsup/jsx": "0.1.4"
+        "whatsup": "^2.2.33"
     },
     "devDependencies": {
-        "@babel/cli": "7.10.4",
-        "@babel/core": "7.10.4",
-        "@babel/plugin-proposal-class-properties": "7.12.1",
-        "@babel/plugin-transform-runtime": "7.10.4",
-        "@babel/plugin-transform-typescript": "7.10.4",
-        "@babel/preset-env": "7.10.4",
-        "@babel/preset-typescript": "7.10.4",
-        "@whatsup/babel-plugin-transform-jsx": "0.2.2",
-        "babel-loader": "8.1.0",
-        "core-js": "3.6.5",
-        "prettier": "2.0.5",
-        "typescript": "3.9.6",
-        "webpack": "4.44.2",
-        "webpack-cli": "3.3.12"
+        "@babel/cli": "7.18.6",
+        "@babel/core": "7.18.6",
+        "@babel/plugin-proposal-class-properties": "7.18.6",
+        "@babel/plugin-proposal-decorators": "^7.18.6",
+        "@babel/plugin-transform-typescript": "7.18.8",
+        "@babel/preset-env": "7.18.6",
+        "@babel/preset-typescript": "7.18.6",
+        "@whatsup/babel-plugin-transform-jsx": "^2.2.33",
+        "babel-loader": "8.2.5",
+        "core-js": "3.23.4",
+        "prettier": "2.7.1",
+        "tslib": "^2.4.0",
+        "typescript": "4.7.4",
+        "webpack": "5.73.0",
+        "webpack-cli": "4.10.0"
     }
 }
