@@ -2,7 +2,7 @@ import React from 'react'
 import { TableResultGeommeanEntry, T_SORT_BY_GEOMMEAN} from '../Common'; 
 
 const GeomMeanRow = ({geomMean, currentSortKey, sortBy, sortbyGeommeanEnum}: {geomMean: Array<TableResultGeommeanEntry|null>; currentSortKey: string;
-    sortBy: (name: string) => void ; sortbyGeommeanEnum: T_SORT_BY_GEOMMEAN;}) => {
+    sortBy: (name: string) => void ; sortbyGeommeanEnum: T_SORT_BY_GEOMMEAN;}): JSX.Element => {
     const sort = (sortValue: string) => (event: React.SyntheticEvent) => {
         event.preventDefault();
         sortBy(sortValue)
