@@ -1,4 +1,4 @@
-import { Browser, Page } from "playwright-core";
+import { Browser, Page, CDPSession } from "playwright-core";
 import { BenchmarkType } from "./benchmarksCommon";
 import { benchmarks, CPUBenchmarkPlaywright, fileNameTrace, MemBenchmarkPlaywright, TBenchmarkPlaywright } from "./benchmarksPlaywright";
 import { BenchmarkOptions, config as defaultConfig, ErrorAndWarning, FrameworkData, TConfig } from "./common";
@@ -11,13 +11,6 @@ let config: TConfig = defaultConfig;
 async function runBenchmark(browser: Browser, page: Page, benchmark: TBenchmarkPlaywright, framework: FrameworkData): Promise<any> {
   await benchmark.run(browser, page, framework);
   if (config.LOG_PROGRESS) console.log("after run ", benchmark.benchmarkInfo.id, benchmark.type, framework.name);
-}
-
-async function afterBenchmark(browser: Browser, page: Page, benchmark: TBenchmarkPlaywright, framework: FrameworkData): Promise<any> {
-  if (benchmark.after) {
-    await benchmark.after(page, framework);
-    if (config.LOG_PROGRESS) console.log("after benchmark ", benchmark.benchmarkInfo.id, benchmark.type, framework.name);
-  }
 }
 
 async function initBenchmark(browser: Browser, page: Page, benchmark: TBenchmarkPlaywright, framework: FrameworkData): Promise<any> {
@@ -53,20 +46,17 @@ function convertError(error: any): string {
   }
 }
 
-// async function forceGC(page: Page) {
-//   const prototypeHandle = await page.evaluateHandle(() => Object.prototype);
-//   const objectsHandle = await page.queryObjects(prototypeHandle);
-//   const numberOfObjects = await page.evaluate((instances) => instances.length, objectsHandle);
-
-//   await Promise.all([prototypeHandle.dispose(), objectsHandle.dispose()]);
-
-//   return numberOfObjects;
-// }
+async function forceGC(page: Page, client: CDPSession) {
+  for (let i=0;i<7;i++) {
+      // await client.send('HeapProfiler.collectGarbage');
+      await page.evaluate('window.gc()');
+  }
+}
 
 async function runCPUBenchmark(framework: FrameworkData, benchmark: CPUBenchmarkPlaywright, benchmarkOptions: BenchmarkOptions): Promise<ErrorAndWarning>
 {
-    let error: String = undefined;
-    let warnings: String[] = [];
+    let error: string = undefined;
+    let warnings: string[] = [];
     let results: number[] = [];
 
     console.log("benchmarking ", framework, benchmark.benchmarkInfo.id);
@@ -81,33 +71,37 @@ async function runCPUBenchmark(framework: FrameworkData, benchmark: CPUBenchmark
                 console.log(`BROWSER: ${msg.args()[i]}`);
             });
         // }
+        let client = await page.context().newCDPSession(page);
         for (let i = 0; i <benchmarkOptions.batchSize; i++) {
             await page.goto(`http://${benchmarkOptions.HOST}:${benchmarkOptions.port}/${framework.uri}/index.html`, {waitUntil: "networkidle"});
 
             console.log("initBenchmark Playwright");
             await initBenchmark(browser, page, benchmark, framework);
-
-            // let categories = [
-            //     "blink.user_timing",
-            //     "devtools.timeline",
-            //     'disabled-by-default-devtools.timeline',
-            // ];
+            // minimal categories
+            // result svelte-v3.46.2-keyed_01_run1k.json min 67.237 max 71.611 mean 69.4036 median 69.20949999999999 stddev 1.3601652023845432
+            // default categories:
+            // result svelte-v3.46.2-keyed_01_run1k.json min 61.334 max 65.92 mean 63.84379999999999 median 63.756 stddev 1.5925086987377977
             let categories = [
-            "loading",
-            'devtools.timeline',
-              'disabled-by-default-devtools.timeline',
-              '-*',
-              'v8.execute',
-                'disabled-by-default-devtools.timeline.frame',
-                'toplevel',
-                'blink.console',
-                'blink.user_timing',
-                'latencyInfo',
-                'disabled-by-default-v8.cpu_profiler',                
-                'disabled-by-default-devtools.timeline.stack',
+                "blink.user_timing",
+                "devtools.timeline",
+                'disabled-by-default-devtools.timeline',
             ];
+            // let categories = [
+            // "loading",
+            // 'devtools.timeline',
+            //   'disabled-by-default-devtools.timeline',
+            //   '-*',
+            //   'v8.execute',
+            //     'disabled-by-default-devtools.timeline.frame',
+            //     'toplevel',
+            //     'blink.console',
+            //     'blink.user_timing',
+            //     'latencyInfo',
+            //     'disabled-by-default-v8.cpu_profiler',                
+            //     'disabled-by-default-devtools.timeline.stack',
+            // ];
 
-            let client = await page.context().newCDPSession(page);
+            await forceGC(page, client);
             if (benchmark.benchmarkInfo.throttleCPU) {
               console.log("CPU slowdown", benchmark.benchmarkInfo.throttleCPU);
               await client.send('Emulation.setCPUThrottlingRate', { rate: benchmark.benchmarkInfo.throttleCPU });
@@ -118,21 +112,13 @@ async function runCPUBenchmark(framework: FrameworkData, benchmark: CPUBenchmark
                 categories:categories
             });
             console.log("runBenchmark Playwright");
-            // let m1 = await page.metrics();
             await runBenchmark(browser, page, benchmark, framework);
 
             await wait(40);
             await browser.stopTracing();
-            // let m2 = await page.metrics();
-            await afterBenchmark(browser, page, benchmark, framework);
             if (benchmark.benchmarkInfo.throttleCPU) {
-              await client.send('Emulation.setCPUThrottlingRate', { rate: 1 });
-            await client.send('HeapProfiler.collectGarbage',{});
-          }
-          client.detach();
-  
-            // console.log("afterBenchmark", m1, m2);
-            // let result = (m2.TaskDuration - m1.TaskDuration)*1000.0; //await computeResultsCPU(fileNameTrace(framework, benchmark, i), benchmarkOptions, framework, benchmark, warnings, benchmarkOptions.batchSize);
+              await client.send('Emulation.setCPUThrottlingRate', { rate: 1 });            
+          }  
             let result = await computeResultsCPU(config, fileNameTrace(framework, benchmark.benchmarkInfo, i), benchmark.benchmarkInfo.durationMeasurementMode);
             results.push(result);
             console.log(`duration for ${framework.name} and ${benchmark.benchmarkInfo.id}: ${result}`);
@@ -168,16 +154,16 @@ async function runMemBenchmark(
   benchmark: MemBenchmarkPlaywright,
   benchmarkOptions: BenchmarkOptions
 ): Promise<ErrorAndWarning> {
-  let error: String = undefined;
-  let warnings: String[] = [];
+  let error: string = undefined;
+  let warnings: string[] = [];
   let results: number[] = [];
 
   console.log("benchmarking ", framework, benchmark.benchmarkInfo.id);
   let browser: Browser = null;
   try {
     browser = await startBrowser(benchmarkOptions);
+    const page = await browser.newPage();
     for (let i = 0; i < benchmarkOptions.batchSize; i++) {
-      const page = await browser.newPage();
       if (config.LOG_DETAILS) {
         page.on("console", (msg) => {
           for (let i = 0; i < msg.args().length; ++i) console.log(`BROWSER: ${msg.args()[i]}`);
@@ -200,18 +186,17 @@ async function runMemBenchmark(
 
       console.log("runBenchmark");
       await runBenchmark(browser, page, benchmark, framework);
+      await forceGC(page, client);
       await wait(40);
-      await client.send('HeapProfiler.collectGarbage');
-      let result = (await client.send('Performance.getMetrics')).metrics.filter((m) => m.name==='JSHeapUsedSize')[0].value / 1024 / 1024;
+      // let result = (await client.send('Performance.getMetrics')).metrics.filter((m) => m.name==='JSHeapUsedSize')[0].value / 1024 / 1024;
 
-      await afterBenchmark(browser, page, benchmark, framework);
+      let result = (await page.evaluate("performance.measureUserAgentSpecificMemory()") as any).bytes / 1024 / 1024;
       console.log("afterBenchmark ");
-      // let result = metrics.metrics.JSHeapUsedSize / 1024.0 / 1024.0;
       results.push(result);
       console.log(`memory result for ${framework.name} and ${benchmark.benchmarkInfo.id}: ${result}`);
       if (result < 0) throw new Error(`memory result ${result} < 0`);
-      await page.close();
     }
+    await page.close();
     await browser.close();
     return { error, warnings, result: results };
   } catch (e) {
