@@ -10,18 +10,11 @@ import { buildDriver, setButtonsInShadowRoot, setShadowRootName, setUseRowShadow
 let config: TConfig = defaultConfig;
 
 // necessary to launch without specifiying a path
-var chromedriver: any = require("chromedriver");
+require("chromedriver");
 
 async function runBenchmark(driver: WebDriver, benchmark: CPUBenchmarkWebdriverCDP, framework: FrameworkData): Promise<any> {
   await benchmark.run(driver, framework);
   if (config.LOG_PROGRESS) console.log("after run ", benchmark.benchmarkInfo.id, benchmark.benchmarkInfo.type, framework.name);
-}
-
-async function afterBenchmark(driver: WebDriver, benchmark: CPUBenchmarkWebdriverCDP, framework: FrameworkData): Promise<any> {
-  if (benchmark.after) {
-    await benchmark.after(driver, framework);
-    if (config.LOG_PROGRESS) console.log("after benchmark ", benchmark.benchmarkInfo.id, benchmark.benchmarkInfo.type, framework.name);
-  }
 }
 
 async function initBenchmark(driver: WebDriver, benchmark: CPUBenchmarkWebdriverCDP, framework: FrameworkData): Promise<any> {
@@ -66,8 +59,8 @@ async function runCPUBenchmark(
   benchmark: CPUBenchmarkWebdriverCDP,
   benchmarkOptions: BenchmarkOptions
 ): Promise<ErrorAndWarning> {
-  let error: String = undefined;
-  let warnings: String[] = [];
+  let error: string = undefined;
+  let warnings: string[] = [];
   let results: number[] = [];
 
   console.log("benchmarking ", framework, benchmark.benchmarkInfo.id, "with webdriver (tracing via CDP Connection)");
@@ -136,8 +129,6 @@ async function runCPUBenchmark(
         console.log("resetting CPU slowdown");
         await (driver as any).sendDevToolsCommand("Emulation.setCPUThrottlingRate", { rate: 1 });
       }
-      await afterBenchmark(driver, benchmark, framework);
-
       await cdpConnection.execute("Tracing.end", {});
       await p;
 
